requirements.txt
# AICO Workspace — Dependencies
# TITAN Bot (execution/titan/)
feedparser
gTTS
python-dotenv
requests
groq
google-genai
pydub
numpy
httpx
